Coding Cleanup
Tried to make the component files more generic. The component methods
should take the input of a pin and return the printed information. For
example EMGprint(A0) takes the A0 pin and prints what it reads from A0
this we we do not need to write a file for each components.